// src/shared/generated/BrowserTabReloadReceiptRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserRoleReloadReceiptRecord } from "./BrowserRoleReloadReceiptRecord";
import type { SystemRuntimeOperationSummaryRecord } from "./SystemRuntimeOperationSummaryRecord";

export type BrowserTabReloadReceiptRecord = { receipt: SystemRuntimeOperationSummaryRecord, roles: Array<BrowserRoleReloadReceiptRecord>, };